refactor: routers folder structure - move middleware.py into sessions utility folder
