+++ backend/server/routers/courses.py
@@ -12,7 +12,7 @@ from algorithms.objects.user import User
 from data.config import ARCHIVED_YEARS
 from fastapi import APIRouter, HTTPException, Security
 from fuzzywuzzy import fuzz  # type: ignore
-from server.routers.auth_utility.middleware import HTTPBearerToUserID
+from server.routers.auth_utility.sessions.middleware import HTTPBearerToUserID
 from server.routers.auth_utility.user import get_setup_user
 from server.db.mongo.conn import archivesDB, coursesCOL
 from server.routers.model import (CACHED_HANDBOOK_NOTE, CONDITIONS, CourseCodes, CourseDetails, CoursesPath,
